Cleaned up obsolete gtest projects and renamed base test project
