--- tests/static_string/pch.cpp
@@ -1,5 +1,0 @@
-//
-// pch.cpp
-//
-
-#include "pch.h"
