--- tests/vld_main_test/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.googletest.v140.windesktop.msvcstl.dyn.rt-dyn" version="1.8.1.7" targetFramework="native" />
-</packages>
