--- tests/vld_main_test/pch.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "gtest/gtest.h"
