+++ tests/vld_tests/Allocs.cpp
