+++ tests/vld_tests/basics.cpp
@@ -4,16 +4,16 @@
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
-namespace msstaticstring
+namespace vld_tests
 {
-    TEST_CLASS(basics_disabled)
+    TEST_CLASS(basics)
     {
-        bool bFreeMemory = false; // todo: this used to be a parameter
+        bool bFreeMemory = true; // todo: this used to be a parameter
     public:
 
         TEST_METHOD_INITIALIZE(Setup)
         {
-            VLDDisable();
+            VLDMarkAllLeaksAsReported();
         }
 
         TEST_METHOD(TestMalloc)
@@ -22,7 +22,8 @@ namespace msstaticstring
             LeakMemoryMalloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 2;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestNew)
@@ -31,7 +32,8 @@ namespace msstaticstring
             LeakMemoryNew(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 2;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestNewArray)
@@ -40,7 +42,8 @@ namespace msstaticstring
             LeakMemoryNewArray(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 2;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestCalloc)
@@ -49,7 +52,8 @@ namespace msstaticstring
             LeakMemoryCalloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 2;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestRealloc)
@@ -58,7 +62,8 @@ namespace msstaticstring
             LeakMemoryRealloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 2;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestCoTaskMem)
@@ -67,7 +72,8 @@ namespace msstaticstring
             LeakMemoryCoTaskMem(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 2;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestAlignedMalloc)
@@ -76,7 +82,8 @@ namespace msstaticstring
             LeakMemoryAlignedMalloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 3;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestAlignedRealloc)
@@ -85,7 +92,8 @@ namespace msstaticstring
             LeakMemoryAlignedRealloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 3;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestStrdup)
@@ -94,7 +102,8 @@ namespace msstaticstring
             LeakMemoryStrdup(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 4;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestHeapAlloc)
@@ -103,7 +112,8 @@ namespace msstaticstring
             LeakMemoryHeapAlloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 1;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestIMalloc)
@@ -112,7 +122,8 @@ namespace msstaticstring
             LeakMemoryIMalloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 1;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD(TestGetProcMalloc)
@@ -121,13 +132,14 @@ namespace msstaticstring
             LeakMemoryGetProcMalloc(repeats, bFreeMemory);
             int total = static_cast<int>(VLDGetLeaksCount());
             int leaks = total - prev;
-            Assert::AreEqual(0, leaks);
+            int correctLeaks = bFreeMemory ? 0 : repeats * 1;
+            Assert::AreEqual(correctLeaks, leaks);
         }
 
         TEST_METHOD_CLEANUP(TearDown)
         {
             // Check that callstack resolved without unresolved functions (required symbols for all dll's)
-            VLDEnable();
+            Assert::AreEqual(0, VLDResolveCallstacks());
         }
     };
 }
