+++ vld.sln
@@ -11,8 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TestHelpers", "TestHelpers"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "corruption", "tests\corruption\corruption.vcxproj", "{528127C3-6892-4D74-868C-B2DAFF72F1E6}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ignore_functions", "tests\ignore_functions\ignore_functions.vcxproj", "{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "vld_dll1", "tests\vld_dll1\vld_dll1.vcxproj", "{158ABCCD-3B90-4AB7-8F88-BA4BF568953D}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "vld_dll2", "tests\vld_dll2\vld_dll2.vcxproj", "{BEAB52AE-7C8E-47FD-AC40-BD15A8296FAE}"
@@ -23,9 +21,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Libs", "Libs", "{CC26D463-0
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "vld_main", "tests\vld_main\vld_main.vcxproj", "{6A5CCF20-1E5D-41B7-A2E2-11F23AA0522C}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ms_static_string", "tests\ms_static_string\ms_static_string.vcxproj", "{C10696B5-E80B-4CE7-8BDF-2C7EC38B4649}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "vld_tests", "tests\vld_tests\vld_tests.vcxproj", "{C10696B5-E80B-4CE7-8BDF-2C7EC38B4649}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "vld_unload", "tests\vld_unload\vld_unload.vcxproj", "{4133697F-1589-4567-9080-19774034DAF9}"
+	ProjectSection(ProjectDependencies) = postProject
+		{158ABCCD-3B90-4AB7-8F88-BA4BF568953D} = {158ABCCD-3B90-4AB7-8F88-BA4BF568953D}
+		{BEAB52AE-7C8E-47FD-AC40-BD15A8296FAE} = {BEAB52AE-7C8E-47FD-AC40-BD15A8296FAE}
+	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "threads", "tests\threads\threads.vcxproj", "{E3A7ACBB-3A20-42F3-87F2-90DA653D7400}"
 EndProject
@@ -63,18 +65,6 @@ Global
 		{528127C3-6892-4D74-868C-B2DAFF72F1E6}.Release|Win32.Build.0 = Release|Win32
 		{528127C3-6892-4D74-868C-B2DAFF72F1E6}.Release|x64.ActiveCfg = Release|x64
 		{528127C3-6892-4D74-868C-B2DAFF72F1E6}.Release|x64.Build.0 = Release|x64
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug(Release)|Win32.ActiveCfg = Debug(Release)|Win32
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug(Release)|Win32.Build.0 = Debug(Release)|Win32
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug(Release)|x64.ActiveCfg = Debug(Release)|x64
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug(Release)|x64.Build.0 = Debug(Release)|x64
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug|Win32.ActiveCfg = Debug|Win32
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug|Win32.Build.0 = Debug|Win32
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug|x64.ActiveCfg = Debug|x64
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Debug|x64.Build.0 = Debug|x64
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Release|Win32.ActiveCfg = Release|Win32
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Release|Win32.Build.0 = Release|Win32
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Release|x64.ActiveCfg = Release|x64
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D}.Release|x64.Build.0 = Release|x64
 		{158ABCCD-3B90-4AB7-8F88-BA4BF568953D}.Debug(Release)|Win32.ActiveCfg = Debug(Release)|Win32
 		{158ABCCD-3B90-4AB7-8F88-BA4BF568953D}.Debug(Release)|Win32.Build.0 = Debug(Release)|Win32
 		{158ABCCD-3B90-4AB7-8F88-BA4BF568953D}.Debug(Release)|x64.ActiveCfg = Debug(Release)|x64
@@ -166,7 +156,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{281D5ACB-9ED2-496B-B19E-A75F4D4DA111} = {9F9CFA3A-F154-4069-89E3-19BDC6BD3A7D}
 		{528127C3-6892-4D74-868C-B2DAFF72F1E6} = {9F9CFA3A-F154-4069-89E3-19BDC6BD3A7D}
-		{9261FDA3-D73D-4B05-A6F4-A28E96765B7D} = {9F9CFA3A-F154-4069-89E3-19BDC6BD3A7D}
 		{158ABCCD-3B90-4AB7-8F88-BA4BF568953D} = {281D5ACB-9ED2-496B-B19E-A75F4D4DA111}
 		{BEAB52AE-7C8E-47FD-AC40-BD15A8296FAE} = {281D5ACB-9ED2-496B-B19E-A75F4D4DA111}
 		{4ADFD279-56C6-4B1D-BA9E-B815E81C1B17} = {CC26D463-0361-4E7E-86E2-BB02AC2ED4ED}
